Completed 2024 day 21
Hardest day so far, took about 15 hours over 2 days to finish.

I had the right idea, but it was a finicky approach (expand each
character with the shortest path every time). Instead, I expand each
character with any of a set of shortest paths and pass it up the chain.
The one that results in the shortest length at the max depth (3 for part
1, 25 for part 2) ends up being picked at each stage.

Part 2 is the same as part 1, but with memoization of the answers at
each depth.

Approach taken from this reddit comment:
https://www.reddit.com/r/adventofcode/comments/1hj2odw/comment/m33lhla/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button

Signed-off-by: Ben Cheatham <Benjamin.Cheatham@amd.com>