Completed 2025 day 11
Completed part 1 on my own.

Came up with the algorithm for part 2 on my own, just couldn't get it
right. I was trying to count the paths from the endpoint to the
startpoint, but I screwed something up. Found an approach that goes in
the forward direction with a memoization map and used that instead.

Signed-off-by: Ben Cheatham <Benjamin.Cheatham@amd.com>